refactor(frontend): Move transformAgreementsJsonBigint to specific module
