+++ src/frontend/src/lib/utils/agreements.utils.ts
@@ -3,27 +3,6 @@ import type { EnvAgreements } from '$env/types/env-agreements';
 import { MILLISECONDS_IN_SECOND } from '$lib/constants/app.constants';
 import { formatSecondsToDate } from '$lib/utils/format.utils';
 
-export const transformAgreementsJsonBigint = (
-	json: Record<string, { lastUpdatedTimestamp: { __bigint__: string }; lastUpdatedDate: string }>
-): EnvAgreements => {
-	const res: Record<string, { lastUpdatedTimestamp: bigint; lastUpdatedDate: string }> = {};
-	Object.entries(json).forEach(
-		([
-			key,
-			{
-				lastUpdatedTimestamp: { __bigint__ },
-				...rest
-			}
-		]) => {
-			res[key] = {
-				...rest,
-				lastUpdatedTimestamp: BigInt(__bigint__)
-			};
-		}
-	);
-	return res as EnvAgreements;
-};
-
 export const getAgreementLastUpdated = ({
 	type,
 	$i18n
